+++ src/Core/ApiClient.php
@@ -37,9 +37,9 @@ class ApiClient implements ApiClientInterface
     /**
      * @const string
      */
-    protected const SDK_VERSION = '2.0.0';
+    protected const SDK_VERSION = '1.1.0';
 
-    protected const SDK_USER_AGENT = 'bitrix24-php-sdk';
+    protected const SDK_USER_AGENT = 'b24-php-sdk';
 
     /**
      * ApiClient constructor.